+++ data/trx/ship/games/tr1/scripts/atlantis.lua
@@ -1,3 +1,3 @@
-trx.events.before_item_setup(function()
+trx.events.on_game_start(function()
   trx.objects.bacon_lara.properties.anchor_room = 10
 end)