+++ data/trx/ship/games/tr2/scripts/floating.lua
@@ -1,4 +1,4 @@
-trx.events.after_level_file(function(level)
+trx.events.on_game_start(function(level)
   trx.objects.swap_mesh(
     trx.catalog.objects.secret_2_option,
     trx.catalog.objects.secret_3_option,