+++ data/trx/ship/games/tr3/scripts/compound.lua
@@ -1,3 +1,3 @@
-trx.events.before_level_file(function(level)
+trx.events.on_game_start(function(level)
   trx.creatures.add_ally(trx.catalog.objects.prisoner)
 end)