+++ data/trx/ship/games/tr3/scripts/nevada.lua
@@ -1,4 +1,4 @@
-trx.events.before_item_setup(function(level)
+trx.events.on_game_start(function(level)
   local props = trx.objects.cobra.properties
   props.alert_radius = 1.25
   props.attack_radius = 0.666667