+++ data/trx/ship/games/tr3/scripts/thames.lua
@@ -1,4 +1,4 @@
-trx.events.before_item_setup(function(level)
+trx.events.on_game_start(function(level)
   trx.objects.propeller_2.properties.damage = 1000
   trx.objects.propeller_3.properties.damage = 1000
 end)