+++ data/trx/ship/games/tr4/scripts/alexhub.lua
@@ -1,4 +1,4 @@
-trx.events.before_item_setup(function(level)
+trx.events.on_game_start(function(level)
   trx.items[38].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
   trx.items[39].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
   trx.items[40].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW