+++ data/trx/ship/games/tr4/scripts/ang_race.lua
@@ -1,9 +1,6 @@
-trx.events.before_item_setup(function(level)
+trx.events.on_game_start(function(level, is_save)
   trx.objects.animating_14.properties.collidable = false
   trx.objects.animating_15.properties.collidable = false
   trx.objects.animating_16.properties.collidable = false
-end)
-
-trx.events.on_game_start(function(level, is_save)
   trx.lara.holsters_visible = trx.lara.has_pistol_weapon
 end)