+++ data/trx/ship/games/tr4/scripts/karnak1.lua
@@ -1,4 +1,4 @@
-trx.events.before_item_setup(function(level)
+trx.events.on_game_start(function(level)
   trx.items[66].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
   trx.items[68].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
   trx.items[82].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW