+++ src/trx/game/level/finalize/gameplay_objects.c
@@ -162,7 +162,6 @@ void Level_Finalize_LoadObjectsAndItems(LEVEL_CONTEXT *const ctx)
     M_PrepareTR4Items(ctx);
 
     Inject_ApplyProperties();
-    LUA_FireEventInt32(LUA_EVENT_BEFORE_ITEM_SETUP, GF_GetCurrentLevel()->num);
 
     const int32_t item_count = Item_GetLevelCount();
     for (int32_t i = 0; i < item_count; i++) {